feat(scrollUP): add the scroll up feature
Added the scroll up feature to the site which helps to navigate the starting point of the site. And added html structure for the feature in the index.html file,
styling in the assets/css/styles.css and the dynamic codes are added in the assets/js/main.js file.