feat(active-link): add the active link feature
Added the active link feature in the nav bar. Which helps the user to navigate which
section of the site is currently viewing. The html structure is added in
the index.html file and styling in the assets/css/styles.css and dynamic
script in the assets/js/main.js